requirements.txt
aiohttp==3.5.4
aiopg==0.16.0
alembic==1.0.11
aiojobs==0.2.2
click==7.0
isort==4.3.21
marshmallow==2.20.1
pyyaml==5.1.2
pytest==5.0.1
pytest-aiohttp==0.3.0
sqlalchemy==1.3.7
